requirements.txt
certifi==2017.4.17
future==0.16.0
python-telegram-bot==6.0.1
wheel==0.24.0
